requirements_new.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
flask-migrate==4.0.5
flask-sqlalchemy==3.0.5

# AI Services
openai==1.3.0
aiohttp==3.9.1

# Security
pyjwt==2.8.0
bcrypt==4.1.1
cryptography==41.0.8

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Logging and Monitoring
structlog==23.2.0

# Async support
asyncio==3.4.3

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0

# Existing dependencies (keep compatibility)
requests==2.31.0
gunicorn==21.2.0
redis==5.0.1